feat: expose the `initializedAos` function
