+++ packages/app-page-builder/src/contexts/PageBuilder/PageElementsProvider.tsx
@@ -5,6 +5,7 @@ import { PageElementsProvider as PbPageElementsProvider } from "@webiny/app-page
 import { createId } from "@webiny/app-page-builder-elements/modifiers/attributes/id";
 import { createClassName } from "@webiny/app-page-builder-elements/modifiers/attributes/className";
 import { createAnimation } from "@webiny/app-page-builder-elements/modifiers/attributes/animation";
+import { initializeAos } from "@webiny/app-page-builder-elements/modifiers/attributes/animation/initializeAos";
 
 // Styles modifiers.
 import { createBackground } from "@webiny/app-page-builder-elements/modifiers/styles/background";
@@ -40,7 +41,7 @@ export const PageElementsProvider: React.FC = ({ children }) => {
         attributes: {
             id: createId(),
             className: createClassName(),
-            animation: createAnimation()
+            animation: createAnimation({ initializeAos })
         },
         styles: {
             background: createBackground(),
